Update to upstream version '7.0.3'
with Debian dir 09f921c1ff1d8030cef8c2d184a0f24fbebd0e9d

+++ rawint.hpp
@@ -173,7 +173,7 @@ inline uint64 GetGreaterOrEqualPow2(uint64 n)
 
 inline uint64 GetLessOrEqualPow2(uint64 n)
 {
-  size_t p=1;
+  uint64 p=1;
   while (p*2<=n)
     p*=2;
   return p;
